scalpX_core/requirements.txt
openai
requests
flask
python-dotenv
gspread
oauth2client
matplotlib
pandas
yfinance
